contracts/contracts/projects/contracts/pyproject.toml
[tool.poetry]
name = "contracts"
version = "0.1.0"
description = "Algorand smart contracts"
authors = ["ABHISHEK <1si23ci002@sit.ac.in>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
algokit-utils = "^4.0.0"
python-dotenv = "^1.0.0"
algorand-python = "^3"
algorand-python-testing = "^1"
hdwallet = "^3.6.1"

[tool.poetry.group.dev.dependencies]
algokit-client-generator = "^2.1.0"
puyapy = "*"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

